modified meal context and removed unwanted stuff
